Add cookbook add_recipe method
